+++ src/styles/themes/accent.js
@@ -1,6 +1,7 @@
 import createMuiTheme from '@material-ui/core/styles/createMuiTheme'
 import memoize from 'lodash/memoize'
-import merge from 'deepmerge'
+import { deepmerge } from '@material-ui/utils'
+
 import accent from '../palettes/accent'
 import { typography, errorStateColours } from './common'
 
@@ -8,7 +9,7 @@ import { typography, errorStateColours } from './common'
  * This theme contains 'accent' coloured swatches
  */
 const theme = createMuiTheme(
-  merge(
+  deepmerge(
     {
       palette: {
         // mui colours