+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Unreleased
 
+* Remove `deepmerge` dependency
+    * We're using MUI's `@material-ui/utils` implementation, added on https://github.com/mui-org/material-ui/pull/17982
+
 # 2.32.0 (2021-05-14)
 
 * Export MUI `Tooltip` and `ClickAwayListener`